application-code/initial-code/admin-backend/requirements.txt
flask
psycopg2-binary
requests
flask-cors
